# [mlir][IR] Turn `FloatType` into a type interface