@matthias-springer 2024-12-06
I wrote a micro benchmark that parses + prints 32768 floats with random floating-point type: https://gist.github.com/matthias-springer/d22cc02b6097553b78bbba39fed8ca1e

Going through the type interface causes a slowdown compared to the hard-coded sequence of `if` checks in `getFloatSemantics`.

```
Benchmark 1: mlir-opt test.mlir -allow-unregistered-dialect
  BEFORE
  Time (mean ± σ):      43.3 ms ±   1.8 ms    [User: 31.9 ms, System: 11.4 ms]
  Range (min … max):    39.8 ms …  48.3 ms    200 runs

  AFTER
  Time (mean ± σ):      50.3 ms ±   1.8 ms    [User: 38.8 ms, System: 11.5 ms]
  Range (min … max):    47.3 ms …  55.1 ms    200 runs
```