@River707 2024-12-16
The immediate changes here look reasonable. Have you done a sweep of the current users of FloatType to see if any of them should be tightened? Opening up to an interface suddenly expands the contract of FloatType.