@matthias-springer 2024-12-16
> The immediate changes here look reasonable. Have you done a sweep of the current users of FloatType to see if any of them should be tightened? Opening up to an interface suddenly expands the contract of FloatType.

From a user perspective this commit is mostly NFC. It does not matter if you have a concrete type or an interface. The API surface is almost identical. There is only one difference that I found: users can no longer attach an interface to `FloatType`, they have to attach it to a concrete float type (see changed test).